requirements.txt
google-cloud-aiplatform
google-cloud-firestore
google-cloud-storage
google-cloud-secret-manager
google-analytics-data
requests
python-dotenv
flask
gunicorn
functions-framework
